lets keep geometry of objects in the canvas in canvas units (double's) :) (and handle incorrect inputs too) :)
SVN revision: 7101

+++ legacy/edje/src/lib/Edje.h
@@ -20,7 +20,7 @@ extern "C" {
    void        *edje_signal_callback_del(Evas_Object *o, const char *emission, const char *source, void (*func) (void *data, Evas_Object *o, const char *emission, const char *source));
    void         edje_signal_emit(Evas_Object *o, const char *emission, const char *source);
 
-   void         edje_part_geometry_get(Evas_Object *o, char *part, int *x, int *y, int *w, int *h);
+   void         edje_part_geometry_get(Evas_Object *o, char *part, double *x, double *y, double *w, double *h);
        
 #ifdef __cplusplus
 }
